refactor: move Log config to Hoard.Types.Environment
