+++ test/Integration/Hoard/DB/PeerPersistenceSpec.hs
@@ -18,6 +18,7 @@ import Hoard.Effects.Log qualified as Log
 import Hoard.Effects.PeerRepo (getPeerByAddress, runPeerRepo, upsertPeers)
 import Hoard.TestHelpers.Database (TestConfig (..), withCleanTestDatabase)
 import Hoard.Types.DBConfig (DBPools (..))
+import Hoard.Types.Environment (defaultLogConfig)
 import Text.Read (read)
 
 
@@ -25,7 +26,7 @@ spec_PeerPersistence :: Spec
 spec_PeerPersistence = do
     let runWrite config action =
             runEff
-                . Log.runLog Log.defaultConfig
+                . Log.runLog defaultLogConfig
                 . runErrorNoCallStack @Text
                 . runDBRead config.pools.readerPool
                 . runDBWrite config.pools.writerPool
